Allow protected methods to be ignored
When you want to extend a global configuration with a configuration for
prod or dev-environment you might need to decalre a formerly private
method with protected visibility. As that up until now required a method
to have a Bean annotation that was actually not possible.

This Commit introduces a method-annotation "@Ignore" to mark protected
methods to be ignored as Bean-methods