Allow overwriting an Alias in an extending class
Currently Disco throws an exception when it finds an already declared
alias. That is totally valid as long as the configuration is done in one
single file.

As soon as you want to have a configuration split over different files
that becomes complicated as you can not overwrite an alias that was
declared in the base class within an extending class.

Given I want a configuration for my production system with an alias for
a logger and I then want to extend that configuration for my
dev-environment and want to implement a different logger.

Currently I need to create three classes. One base class with all the
configs that will never change and one class for production config
extending the base class and implementing all the aliases that will
change between prod and dev. And I also create another config for the
dev-counterpart of those changing aliases.

This commit allows to overwrite aliases in extending files only. So I
can create my configuration for production and for dev I can extend that
base class and overwrite the aliases I need to adapt.

This change should be absolute BC, no tests where adapted, only added.